feat: Address open PRs and issues, update project structure
