+++ .gitignore
@@ -1,2 +1,5 @@
 node_modules
 .DS_Store
+dist
+coverage
+*.log
