--- .prettierrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  trailingComma: 'es5',
-  tabWidth: 2,
-  semi: true,
-  singleQuote: true,
-};
