+++ example.js
@@ -1,6 +1,4 @@
-var jsbn = require('jsbn');
-(function() {
-  var BigInteger = jsbn.BigInteger;
-  var a = new BigInteger('91823918239182398123');
-  console.log(a.bitLength());
-})();
+import { BigInteger } from 'jsbn';
+
+const a = new BigInteger('91823918239182398123');
+console.log(a.bitLength());
